test_task/requirements.txt
amqp==5.2.0
asgiref==3.8.1
better-profanity==0.7.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
Django==5.1.2
django-celery-results==2.5.1
django-environ==0.11.2
django-environment==0.3.8
django_db==0.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
joblib==1.4.2
kombu==5.4.2
Markdown==3.7
numpy==2.1.2
pbr==6.1.0
platformdirs==4.3.6
profanity-check==1.0.3
prompt_toolkit==3.0.48
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
redis==5.1.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
stevedore==5.3.0
threadpoolctl==3.5.0
tzdata==2024.2
vine==5.1.0
virtualenv==20.27.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wcwidth==0.2.13
